move json tests to separate file
